Rename syncserver.py to server.py and syncwalker.py to walker.py
